% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilansiva-class.R
\docType{class}
\name{bilansiva-class}
\alias{bilansiva-class}
\title{Class bilansiva}
\description{
Cette classe utilise tablesiva pour récupérer les infos sur plusieurs tables
}
\section{Slots}{

\describe{
\item{\code{tables}}{Noms des la tables a aller chercher.}

\item{\code{noms}}{Noms des la colonnes.}

\item{\code{daterondes}}{Traitement par la methode datesrondes, soit "constant" pour ouvertures en tout
ou rien des vannes, soit "linear" pour des données continues, soit "none" pour pas de traitement.}

\item{\code{tags}}{Si plusieurs variables sont dans la même table il faut donner leur
tag pour les différentier.}

\item{\code{debut}}{Date de début.}

\item{\code{fin}}{Date de fin.}

\item{\code{bilandata}}{Un data frame avec les données de bilan chargé.}
}}

\note{
Cette classe teste que les longueurs des champs nom, daterondes,et table sont égales
}
