% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_fonct_passe.R
\name{calcul_fonct_pass}
\alias{calcul_fonct_pass}
\title{Calcul du fonctionnement de la passe d'Arzal}
\usage{
calcul_fonct_pass(dat, per_dis_identifiant = 1)
}
\arguments{
\item{dat}{jeu de données contenant les colonnes c("niveau_vilaine_passe","niveau_mer_passe","niveau_vilaine_barrage","niveau_mer_barrage","niveau_sas_passe",
"hauteur_vanne_vilaine","hauteur_vanne_mer")}

\item{per_dis_identifiant}{Identifiant du dispositif de franchissement, Default: 1}
}
\value{
list with dat=data frame avec les données corrigées et les fonctionnements de la passe (types d'arrêts),
t_periodefonctionnement_dis le tableau à réimporter dans stacomi contenant les périodes de fonctionnement des dispositifs
}
\description{
Par défaut la fonction utilise les niveaux mer et Vilaine barrage puis les niveaux passe
quand ces derniers sont manquants. Calcule les périodes continues avec les mêmes paramètres pour per_tar_code
à l'aide de la fonction rle
}
\examples{
\dontrun{
if(interactive()){
 #dat <- data_DF_init
 #calc_fonct_pass(dat)
 }
}
}
