% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_debit_passe.R
\name{calcul_debit_passe}
\alias{calcul_debit_passe}
\title{Calcul du débit de la passe}
\usage{
calcul_debit_passe(list_fonct_pass)
}
\arguments{
\item{list_fonct_pass}{Une liste contenant un data.frame dat détaillant
le foncionnement de la passe par pas de temps de 10 minutes (voir fonction calc_fonct_passe)}
}
\value{
list_fonct_pass avec calcul du débit dans la passe
}
\description{
Calcule le débit de la passe à partir des types de fonctionnement du DF et du coefficient de
noyage de la vanne aval
}
\examples{
\dontrun{
if(interactive()){
 #calcul_debit_passe(ll)
 }
}
}
