% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2001.R
\docType{data}
\name{dat2001}
\alias{dat2001}
\title{Jeu de données de 2001 pour le calcul de débit}
\format{
\subsection{\code{dat2001}}{

A data frame with 2338 rows and 28 columns:
\describe{
\item{HoroDate}{Horodate}
\item{volet1,...}{position du volet}
\item{vanne1,...}{position de la vanne}
...
}
}
}
\usage{
dat2001
}
\description{
généré par load_debit_barrage
}
\keyword{datasets}
