% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2003.R
\docType{data}
\name{dat2003}
\alias{dat2003}
\title{Jeu de données de 2003 pour le calcul de débit}
\format{
\subsection{\code{dat2003}}{

A data frame with 52558 rows and 26 columns:
\describe{
\item{HoroDate}{Horodate}
\item{volet1,...}{position du volet}
\item{vanne1,...}{position de la vanne}
...
}
}
}
\usage{
dat2003
}
\description{
généré par load_debit_barrage
}
\keyword{datasets}
