% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2019.R
\docType{data}
\name{dat2019}
\alias{dat2019}
\title{Jeu de données de 2019 pour le calcul de débit}
\format{
\subsection{\code{dat2019}}{

A data frame with 52558 rows and 52 columns:
\describe{
\item{horodate}{Horodate}
\item{volet1,...}{position du volet}
\item{vanne1,...}{position de la vanne}
...
}
}
}
\usage{
dat2019
}
\description{
généré par load_debit_barrage
}
\keyword{datasets}
