% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datesrondes-tablesiva.R
\name{datesrondes,tablesiva-method}
\alias{datesrondes,tablesiva-method}
\title{Methode daterondes}
\usage{
\S4method{datesrondes}{tablesiva}(objet, method = "constant")
}
\arguments{
\item{objet}{Un objet de classe tablesiva}

\item{method, }{méthode d'interpolation linéaire passée à approx, "linear" ou "constant" ou "none".}
}
\description{
La méthode datesrondes permet d'arrondir les dates, les données sont passées de rawdata à corrdata.
Deux options methode="constant" et method="linear" voir approx()
}
