% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debit.R
\docType{data}
\name{debit}
\alias{debit}
\title{Variables de debit dans SIVA}
\format{
\subsection{\code{variable}}{

Un data frame avec 9 lignes et 9 colonnes:
\describe{
\item{code}{Code interne pour appel de la variable}
\item{tag}{Le tag}
\item{TagStation}{Le tag de la station}
\item{Libelle}{Le libelle modifié depuis SIVA}
\item{unite}{L'unite}
...
}
}
}
\usage{
debit
}
\description{
extrait depuis SIVA_R voir data_raw variable.R
}
\keyword{datasets}
