% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debit_10_min.R
\name{debit_10_min}
\alias{debit_10_min}
\title{Débit du barrage.}
\usage{
debit_10_min(debit_barrage)
}
\arguments{
\item{debit_barrage}{Le jeu de données issu du chargement \code{load_debit_barrage}.}
}
\value{
Un tableau de données de débit journier.
Un tibble avec :
debit_barQ Le débit source barrage à partir des volumes écluse, passe,
siphon et #' des débits mesurés sur les vannes et les volets
debit_recalculé Le débit recalculé y compris à partir des volumes écluse,
passe, siphon
}
\description{
Attention la valeur du jour 1 est enlevée à cause des totaliseurs.
}
