% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debit_journalier.R
\name{debit_journalier}
\alias{debit_journalier}
\title{Débit journalier du barrage.}
\usage{
debit_journalier(
  debit_barrage,
  type = c("recalcule", "barrage_volume", "barrage_debit")
)
}
\arguments{
\item{debit_barrage}{Le jeu de données issu du chargement \code{load_debit_barrage}.}

\item{type}{Le type de sortie, soit "recacule", sinon "barrage_volume" ou
barrage_debit.}
}
\value{
Un tableau de données de débit journier.
Un tibble avec :

si type = "recacule" :
vol_recalcule : le volume recaculé (m3) par les fonctions de débit de SIVA
pour les vannes et les volets, plus le volume des siphons, de la passe,
et de l'écluse, issus de la base de données du barrage.
debit_moyen_recalcule = vol_recalc / 86400,

si type = barrage_volume :
vol_bar = volume évacué au barrage,
vol_vanne = volume évacué par les vannes,
vol_volet = id... volets,
vol_siphon = id... siphon,
vol_ecluse = id... ecluse,
vol_passe = id... passe,
debit_moyen_vol_bar = vol_bar/86400,
...

si type = barrage_debit :
vol_vanne_barQ = vol vannes à partir des débits,
vol_volet_barQ = vol volet à partir des débits,
debit_vilaine_estime = ramené au jour, on ne sait pas trop ce que c'est,
debit_moyen_cran = débit moyen de Cran calculé comme somme des volumnes
instantannés, divisés par 86400,
volume_total_barQ = volume total à partir des totaliseurs pour passe, ecluse,
siphon, et à partir du débit moyen 10 min des vannes et des volets,
debit_volet_barQ = à partir du débit moyen 10 min  volets,
debit_vanne_barQ = à partir du débit moyen 10 min  vannes,
debit_barQ = volume_total_barQ / 86400.
}
\description{
Attention la valeur du jour 1 est enlevée à cause des totaliseurs.
}
