% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debit_total.R
\name{debit_total}
\alias{debit_total}
\title{Débit de l'ensemble des volets et des vannes du barrage.}
\usage{
debit_total(
  param = c(Cvo = 0.392, Cvg = 1.5, Cvgs = 1.5, Cvw = 1.5),
  param0,
  dat
)
}
\arguments{
\item{param}{Les paramètres à optimiser (4 paramètres "Cvo","Cvg","Cvgs","Cvw") voir}

\item{param0}{}

\item{dat}{}
}
\value{
Un data frame avec
date = la date (au format date),
Qva12345 = la somme des débits des vannes 1 à 5,
Qvo12345 = la somme des débits des volets 1 à 5,
tot_vol_ecluse = le volume total de l'écluse, tot_vol_siphon = le volume total du siphon,
tot_vol_passe = le volume total de la passe,
debit_moyen_cran = le débit moyen journalier au pont de Cran
}
\description{
Débit total (cette fonction retourne le débit total journalier et la valeur moyenne
Pondérée du débit de Cran.) en utilisant les choix IFSW pour les fonctions de débit
Briand et Woimant (2015). Le calcul se fait en utilisant canal="ifsw" et loi_orificenoye = "ifws".
}
\note{
Cette fonction a été utilisée pour optimiser les coefficients de débit,
param est consititué des coefficients optimisés
mais il faut les autres qui sont stocké dans param0 pour faire tourner débit_total
Ainsi a_optimiser tourne bien sur param mais débit total sur param + les coeff de param0.
CHANGEMENT 2016 la référence devient niveaumerb et niveauVilaineb.
}
