% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debit_vanne.R
\name{debit_vanne}
\alias{debit_vanne}
\title{Fonction de calcul du débit d'une vanne au barrage d'Arzal}
\usage{
debit_vanne(
  hvilaine,
  hmer,
  hvanne,
  l = 18,
  g = 9.81,
  canal = "ifsw",
  dm = 73,
  dv = 94,
  Cma = 0.65,
  Cho = 0.79,
  m_oninf4 = 0.663,
  m_onsup4 = 0.995,
  Cvw = 1,
  Cvg = 1.5,
  Cvgs = 1.5,
  loi_orificenoye = "ifws"
)
}
\arguments{
\item{hvilaine}{Vecteur de données des niveaux Vilaine.}

\item{hmer}{Vecteur de données des niveaux mer.}

\item{hvanne}{Vecteur de données des hauteurs de vanne.}

\item{l}{Largeur de la vanne, défaut 18.}

\item{g}{9.81.}

\item{canal}{Loi de débit choisir "bazin", "aubuisson", "horton", "manning" ou "ifsw".}

\item{dm}{Distance à la sonde amont pour formle de Basin défaut 73.}

\item{dv}{Distance à la sonde aval pour formule de Bazin défaut 94.}

\item{Cma}{Coefficient canal pour Manning défaut 0.65.}

\item{Cho}{Coefficient canal pour Horton défaut .79.}

\item{m_oninf4}{Coefficient Orifice noyé noyé hvanne<4 (Ferrette) défaut 0.663.}

\item{m_onsup4}{Orifice noyé hvanne >4 défaut 0.995.}

\item{Cvw}{Coefficient ifsw canal défaut 1.}

\item{Cvg}{Coefficient ifsw orifice h1<=1.5 hvanne défaut 1.5.}

\item{Cvgs}{Coefficient ifsw h1>1.5 hvanne défaut 1.5.}

\item{loi_orificenoye}{défaut "ifws", choisir "ferrette" ou "ifws".}
}
\value{
Un tableau de données avec deux colonnes,
"Q" le débit calculé en m3/s et "typecalc"
le type de calcul "vanne fermee", "hmer>hvilaine", "canal aubuisson",
"canal horton", "canal bazin",
"canal manning", "canal ifsw (noye)", "canal ifsw (libre)",
"orifice noye (ferr)", "orifice noye (ifws) inf1.5" ou
"orifice noye (ifws) sup1.5".
}
\description{
Fonction de calcul de débit pour une vanne
Choisir canal="bazin" pour avoir les calculs au barrage
Aubuisson et horton pour calculer les débits suivant Ferrette (2004) en prenant ou non compte des vitesses
Par rapport à Ferette (2004), la formule en orifice dénoyé calcule les vitesses par iteration.
Le calcul est effectué à partir d'une indexation des conditions de débits (vecteur de valeurs vrai et faux) entre crochet.
Les coefficients de Bazin ont été recalculés par une 'regression' loess sur la base du rapport annexé à Briand et Woimant (2015)
}
\examples{
# Exploration des coefficients la formule de Bazin 

H=c(5:11)
Rh=c(2,2.25,2.5,2.8,3,3.2,3.4)
K=c(118,125,133,140,146,152,156)
plot(Rh,K)
points(Rh,predict(lm(K~Rh)),type="l")
gamma=87*Rh/(K+sqrt(Rh))
plot(H,gamma,pch="A")
df<-data.frame(H=H,gamma=gamma)
loess_gamma_bazin<-loess(gamma~H,data=df,control = 
loess.control(surface = "direct")) 
# contol pour extrapolation en dehors de la gamme
#save(loess_gamma_bazin,file=str_c("inst","loess_gamma_bazin.Rdata"))
df2<-data.frame(H=seq(from=5,to=11,by=0.1)) # pour les prédictions
df2$gamma<-predict(loess_gamma_bazin,newdata=df2)
points(df2$H,df2$gamma,col="red")

# utilisation de la fonction de débit pour une vanne
# chargement des paramètres et des données de la base
load(system.file("param2012_2014.Rdata", package = "SIVA"))
load(system.file("dat2019.Rdata", package = "SIVA"))
# Calcul du débit d'une vanne
Qva1 <-
  debit_vanne(
    hvilaine = dat$niveauvilaineb,
    hmer = dat$niveaumerb,
    hvanne = dat$vanne1,
    canal = "ifsw",
    Cvg = param["Cvg"],
    Cvgs = param["Cvgs"],
    Cvw = param["Cvw"],
    loi_orificenoye = "ifws"
  )
plot(Qva1[4000:5000,])


}
\author{
Cedric Briand \email{cedric.briand@eaux-et-vilaine.bzh},
Stephanie Woimant #' \email{stephanie.woimant@eaux-et-vilaine.bzh}
}
