% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debit_vannes_isac.R
\name{debit_vannes_isac}
\alias{debit_vannes_isac}
\title{Fonction de calcul du débit d'une vanne sur le barrage de l'Isac}
\usage{
debit_vannes_isac(
  hamont,
  haval,
  hvanne1,
  hvanne2,
  lvanne2 = 4.98,
  lvanne1 = 4.81,
  g = 9.81,
  canal = "manning",
  Cma = 0.65,
  hradier = 0.28
)
}
\arguments{
\item{hamont}{Vecteur de données des niveaux amont.}

\item{haval}{Vecteur de données des niveaux aval.}

\item{hvanne1}{Vecteur de données des hauteurs de vanne 1.}

\item{hvanne2}{Vecteur de données des hauteurs de vanne 2.}

\item{lvanne2}{Largeur de la vanne2 en rive, 4.98}

\item{lvanne1}{Larguer de la vanne1 au centre de l'ouvrage 4.81}

\item{g}{9.81.}

\item{canal}{Loi de débit "manning".}

\item{Cma}{Coefficient canal pour Manning défaut 0.65.}
}
\value{
Un data frame avec "Q1" le débit calculé sur la vanne 1 en m3/s,
"Q2" le débit calculé sur la vanne 1 en m3/s, Q le débit total typecalc1,
et typecalc2 le type de calcul, "vanne fermee", "pas de données", "Manning",
"Manning vers marais" cette dernière correspondant à un écoulement depuis la Vilaine
vers le marais, "mig" migration au droit de l'ouvrage (bloquee, transparente, difficile)
}
\description{
Fonction de calcul de débit pour une vanne
}
\examples{
# Calcul du débit d'une vanne
Qva1 <-
  debit_vannes_isac(
    hamont = dat$isac_amont2,
    haval = dat$pontdecran,
    hvanne1 = dat$isac_position_vanne_1,
    hvanne2 = dat$isac_position_vanne_2
  )
plot(Qva1[4000:5000,])


}
\author{
Cedric Briand \email{cedric.briand@eaux-et-vilaine.bzh}
}
