% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debit_volet.R
\name{debit_volet}
\alias{debit_volet}
\title{Débit d'un volet}
\usage{
debit_volet(hvolet, hvilaine, hvanne, hmer, Cvo = 0.627, g = 9.81, l = 16.4)
}
\arguments{
\item{hvolet}{La hauteur du volet.}

\item{hvilaine}{Le niveau Vilaine.}

\item{hvanne}{La hauteur de vanne.}

\item{hmer}{Le niveau mer.}

\item{Cvo}{Le coefficient de débit sur le volet défaut 0.627.}

\item{g}{La pesanteur, toujours 9.81.}

\item{l}{La largeur d'une vanne 16.4}
}
\value{
Le calcul de débit d'un volet Q
}
\description{
Formule déversoir à crête épaisse (Ferrette, 2004).
}
\examples{
load(system.file("param2012_2014.Rdata", package = "SIVA"))
load(system.file("dat2019.Rdata", package = "SIVA"))
dv <-debit_volet(
  hvolet = dat$volet1,
  hvilaine = dat$niveauvilaine,
  hvanne = dat$vanne1,
  hmer = dat$niveaumer,
  Cvo = param["Cvo"]
)
}
\author{
Cedric Briand \email{cedric.briand@eaux-et-vilaine.bzh}, Stephanie Woimant
\email{stephanie.woimant@eaux-et-vilaine.bzh}
}
