% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debit-de-l-ensemble-des-volets-debit_volets.R
\name{debit_volets}
\alias{debit_volets}
\title{Calcul du débit sur tous les volets}
\usage{
debit_volets(
  hvolet1,
  hvolet2,
  hvolet3,
  hvolet4,
  hvolet5,
  hvilaine,
  hvanne1,
  hvanne2,
  hvanne3,
  hvanne4,
  hvanne5,
  hmer,
  Cvo = 0.392
)
}
\description{
Somme du débit de tous les volets
}
