% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicateur_avifaune_isac.R
\name{indicateur_avifaune_isac}
\alias{indicateur_avifaune_isac}
\title{Analyse de l'indicateur avifaune sur l'Isac}
\usage{
indicateur_avifaune_isac(
  dat,
  h_favorable = 2.4,
  h_moyen = 2.2,
  mois_debut = "11",
  mois_fin = "4",
  niveau_marais = "isac_niveau_marais"
)
}
\arguments{
\item{dat}{Un tableau de données, chargé par load_isac et traitement_debit_isac}

\item{h_favorable}{Le niveau favorable à l'accueil d'oiseaux hivernants sur
l'Isac}

\item{h_moyen}{Le niveau en dessous duquel l'accueil des oiseaux hivernants
devient mauvais}

\item{mois_debut}{Le mois de début de l'accueil de l'avifaune (inclus)}

\item{mois_fin}{Le mois de fin de l'accueil de l'avifaune (inclus)}

\item{niveau_marais}{Le niveau dans les marais, defaut "isac_niveau_marais"
voir le jeu de données isac pour les codes possibles}
}
\value{
Ub vecteur (factor) des types pour l'avifaune
}
\description{
Analyse de l'indicateur avifaune sur l'Isac
}
