% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicateur_brochet_isac.R
\name{indicateur_brochet_isac}
\alias{indicateur_brochet_isac}
\title{Analyse de l'indicateur brochet sur l'Isac}
\usage{
indicateur_brochet_isac(
  dat,
  h_favorable = 2.3,
  h_moyen = 2.1,
  mois_debut = "2",
  mois_fin = "4",
  niveau_marais = "isac_niveau_marais"
)
}
\arguments{
\item{dat}{Un tableau de données, chargé par load_isac et traitement_debit_isac}

\item{h_favorable}{Le niveau favorable pour la reproduction}

\item{mois_debut}{Le mois avant lequel la repro s'effectue et après lequel l'émergence commence}

\item{mois_fin}{Le mois de fin de l'émergence (inclus)}

\item{niveau_marais}{Le niveau dans les marais, defaut "isac_niveau_marais"
voir le jeu de données isac pour les codes possibles}
}
\value{
Ub vecteur (factor) des types pour le brochet
}
\description{
note :si on est pas dans la bonne période, pas de calcul du niveau ponte, on
prend 2.30 comme référence....)
}
