% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_into.R
\name{insert_into}
\alias{insert_into}
\title{Fonction pour insérer des données dans SIVA. Attention zone de danger !}
\usage{
insert_into(data, columns, tabledest, baseODBC, ...)
}
\arguments{
\item{data}{Les données à exporter.}

\item{columns}{Le nom des colonnes à exporter.}

\item{tabledest}{Le nom de la table de destination dans la base.}

\item{baseODBC}{Le lien ODBC vers la base.}

\item{...}{Autres paramètres passés à la fonction.}
}
\value{
nothing
}
\description{
Fonction pour insérer des données dans SIVA. Attention zone de danger !
}
\examples{
#TODO
}
