% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isac_dat.R
\docType{data}
\name{isac_dat}
\alias{isac_dat}
\title{Jeu de données de 2020 pour l'Isac}
\format{
\subsection{\code{isac_dat}}{

A data frame with 52588 rows and 9 columns:
\describe{
\item{horodate}{Horodate}
\item{pontdecran,...}{Niveau au pont de cran}
\item{...}{autres variables, voir \code{isac}  et data_raw/isac_dat.R pour les variables chargées}
...
}
}
}
\usage{
isac_dat
}
\description{
généré par load_isac
}
\keyword{datasets}
