% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_debit_barrage.R
\name{load_debit_barrage}
\alias{load_debit_barrage}
\title{Chargement des donnees de debit du barrage}
\usage{
load_debit_barrage(debut, fin, con)
}
\arguments{
\item{debut}{La date de début format POSIXct ou character "\%Y-\%m-\%d \%H:\%M:\%S}

\item{fin}{La fin format POSIXct ou character "\%Y-\%m-\%d \%H:\%M:\%S}

\item{con}{une connexion pool}
}
\value{
Le tableau de données pour le calcul des debits du barrage
}
\description{
Il faut avoir configuré la connexion à la base
}
\examples{
\dontrun{
load_debit_barrage(debut = 
as.POSIXct(strptime("2010-01-01 00:00:00", format = "\%Y-\%m-\%d \%H:\%M:\%S")),
   fin = 
   as.POSIXct(strptime("2010-01-10 00:00:00", format = "\%Y-\%m-\%d \%H:\%M:\%S")),
   con= con)
   }
}
