% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_vega.R
\name{load_vega}
\alias{load_vega}
\title{Collecte les valeurs de l'API vega}
\usage{
load_vega(
  sonde = "Trevelo",
  horodate_debut = "2023-01-01 00:00",
  horodate_fin = NULL,
  apiKey
)
}
\arguments{
\item{sonde}{Soit "Trevelo" soit "Isac"}

\item{horodate_debut}{La date de début au format "2023-01-01 00:00"}

\item{horodate_fin}{La date de début au format "2023-01-01 00:00", défaut NULL,
correspond alors à la date du jour}

\item{apiKey}{La clé de l'API}
}
\value{
Un tableau de données avec "timestamp","value","state","isvalid"
}
\description{
Collecte les valeurs de l'API vega
}
\examples{
 \dontrun{
 load_vega("Trevelo")
 load_vega("Isac")
 load_vega("Isac", horodate_debut="2023-01-01 00:00")
 }
}
