% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddb-tablesiva.R
\name{loaddb,tablesiva-method}
\alias{loaddb,tablesiva-method}
\title{Methode loaddb}
\usage{
\S4method{loaddb}{tablesiva}(objet, con, checktag = FALSE)
}
\arguments{
\item{objet, }{un objet de classe tablesiva}

\item{con, }{une connexion pool}

\item{checktag, }{faut il faire un check de l'existence d'un tag dans la table ?}
}
\value{
Un object de classe tablesiva
}
\description{
La méthode loaddb permet d'aller chercher les données
}
