% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niveau.R
\docType{data}
\name{niveau}
\alias{niveau}
\title{Variables de niveau dans SIVA}
\format{
\subsection{\code{variable}}{

Un data frame avec 33 lignes et 7 colonnes:
\describe{
\item{code}{Code interne pour appel de la variable}
\item{tag}{Le tag}
\item{TagStation}{Le tag de la station}
\item{Libelle}{Le libelle modifié depuis SIVA}
\item{unite}{L'unite}
...
}
}
}
\usage{
niveau
}
\description{
extrait depuis SIVA_R voir data_raw variable.R
}
\keyword{datasets}
