% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\docType{data}
\name{param}
\alias{param}
\title{Paramètres calés pour le calcul de débit au barrage d'Arzal (voir Briand et Woimant 2015)}
\format{
\subsection{\code{param}}{

named vector"Cvo" "Cvg" "Cvgs" "Cvw"
\describe{
\item{Cvo}{Coeff volet}
\item{Cvg}{coefficient ifsw orifice h1<=1.5 hvanne}
\item{Cvgs}{coefficient ifsw h1>1.5 hvanne}
\item{Cvw}{coefficient ifsw canal}
...
}
}
}
\usage{
param
}
\description{
Paramètres calés pour le calcul de débit au barrage d'Arzal (voir Briand et Woimant 2015)
}
\keyword{datasets}
