% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_debit_passe.R
\name{plot_debit_passe}
\alias{plot_debit_passe}
\title{Plot avec le débit passe barrage et le débit recalculé}
\usage{
plot_debit_passe(ll)
}
\arguments{
\item{ll}{la liste retournée par calc_fonct_passe}
}
\description{
retourne un ggplot
}
