% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_passe_detail.R
\name{plot_passe_detail}
\alias{plot_passe_detail}
\title{Plot avec lignes et indications du fonctionnement du DF}
\usage{
plot_passe_detail(data_DF, debut, fin)
}
\arguments{
\item{data_DF}{le tableau des niveaux du df
avec une date de début une date de fin et un changement à chaque étape.}

\item{debut}{date de début, character or date}

\item{fin}{date de fin}
}
\value{
Un graphique ggplot
}
\description{
retourne un ggplot
}
\examples{
\dontrun{
if(interactive()){
 #plot_passe_detail(ll$dat)
 }
}
}
