% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sql.R
\name{prepare_sql}
\alias{prepare_sql}
\title{Prepare les données pour l'insertion (cast au bon format)}
\usage{
prepare_sql(data)
}
\arguments{
\item{data}{Un data.frame.}
}
\value{
data Un data.frame.
}
\description{
Prepare les données pour l'insertion (cast au bon format)
}
\examples{
#TODO
}
