% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata2020.R
\docType{data}
\name{rawdata2020}
\alias{rawdata2020}
\title{Jeu de données court de 2020 montrant l'usage de load_debit_barrage sans traitement supplémentaire}
\format{
\subsection{\code{rawdata2020}}{

A data frame with 2735 rows and 35 columns:
\describe{
\item{horodate}{Horodate}
\item{volet1,...}{position du volet}
\item{vanne1,...}{position de la vanne}
...
}
}
}
\usage{
rawdata2020
}
\description{
généré par load_debit_barrage
}
\keyword{datasets}
