% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siva-package.R
\docType{package}
\name{SIVA-package}
\alias{SIVA}
\alias{SIVA-package}
\title{SIVA: Systeme D'information Sur La Vilaine Aval}
\description{
Interface graphique vers SIVA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Eaux-et-Vilaine/SIVA}
  \item Report bugs at \url{https://github.com/Eaux-et-Vilaine/SIVA/issues}
}

}
\author{
\strong{Maintainer}: Cédric Briand \email{cedric.briand@eaux-et-vilaine.bzh} (\href{https://orcid.org/0000-0002-7721-5195}{ORCID})

Authors:
\itemize{
  \item Anthony De Burghrave \email{anthony.deburghrave@eaux-et-vilaine.bzh}
}

}
\keyword{internal}
