% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivaplot-tablesiva.R
\name{sivaplot,tablesiva-method}
\alias{sivaplot,tablesiva-method}
\title{Methode sivaplot}
\usage{
\S4method{sivaplot}{tablesiva}(objet)
}
\arguments{
\item{objet, }{un objet de classe tablesiva}
}
\value{
Un object de classe tablesiva
}
\description{
La méthode sivaplot affiche les données corrigées et les données brutes (avec un décalage de date)
}
