% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablesiva-class.R
\docType{class}
\name{tablesiva-class}
\alias{tablesiva-class}
\title{Classe de connexion SIVA}
\value{
Un objet de classe tablesiva.
}
\description{
Classe de connexion SIVA
}
\section{Slots}{

\describe{
\item{\code{table}}{Nom de la table (serie de données) dans SIVA.}

\item{\code{nom}}{Nom de la série dans le tableau de données.}

\item{\code{tag}}{Tag dans siva}

\item{\code{debut}}{Horodate de début.}

\item{\code{fin}}{Horodate de fin.}

\item{\code{rawdata}}{Données brutes.}

\item{\code{corrdata}}{Données corrigées.}

\item{\code{con}}{Une connexion de classe stacomirtools}
}}

