% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement_cumul.R
\name{traitement_cumul}
\alias{traitement_cumul}
\title{Traitement des données par exemple en sortie de load_isac}
\usage{
traitement_cumul(dat, pattern = "cumul")
}
\arguments{
\item{dat}{Un tableau de données, les données avec cumul doivent avoir
pattern dans leur nom}

\item{pattern}{La chaine de caractère identifiant les cumuls, défaut "cumul"}
}
\value{
Un tableau de données avec diff pour les totaux
}
\description{
traitement des cumuls
}
\note{
Pour les données de débits utiliser la fonction traitement_SIVA
}
\seealso{
traitement_SIVA pour les données de débit du barrage.
}
