% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement_isac.R
\name{traitement_isac}
\alias{traitement_isac}
\title{Traitement des données de l'isac}
\usage{
traitement_isac(
  dat,
  cout_elec = 0.174,
  puissance_moy_pompe = 50,
  cout_carbone_kwh = 0.1
)
}
\arguments{
\item{dat}{Un tableau de données de l'isac chargées avec \code{\link{load_isac}}}

\item{cout_elec}{Cout de l'electricité Kwh defaut 0.1740}

\item{puissance_moy_pompe}{Puissance moyenne de la pompe de l'Isac defaut 50 KW}

\item{cout_carbone_kwh}{Cout en kg de CO2 du kilowatt here défaut 0.1}
}
\value{
Un tableau de données avec les donnees d'entrée et
\begin{itemize}
\item{h}{le delta amont aval en m (les débits pompe sont calculés avec -h).}
\item{t}{ le temps de fonctionnement cumulé des pompes ... corrigé en enlevant
les fonctionnements vannes ouvertes et les valeurs > 40 min par 10 min.}
\item{cout_euros}{ Le coût en euros de fonctionnement (juste la conso électrique, ne prend pas
en compte de le coût de l'abonnement).}
\item{cout_carbone}{ Le coût carbone de fonctionnement des pompes basé sur le cout carbone
du KWh en France.}
\end{itemize}
}
\description{
traitement des cumuls + correction
}
\note{
Les données sont corrigées comme suit :
les données avec des données de débit pompe >0 et vanne ouvertes =>0,
les données avec des débits horaires supérieur à 40 sont mis en NA
la consommation électrique des pompes est calculée à partir de la puissance
moyenne 50 kW. La conso carbone en utilisant un coefficient moyen.
}
