% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement_siva.R
\name{traitement_siva}
\alias{traitement_siva}
\title{Traitement des données en sortie de SIVA}
\usage{
traitement_siva(dat)
}
\arguments{
\item{dat}{Un tableau de données cont}
}
\value{
Un tableau de données propres
}
\description{
recalcule des données des compteurs, enlève les valeurs abérrantes des débits
des siphons
}
\examples{
# voir example-bilansiva-debit pour le chargement des données de 2020 dans SIVA
load(system.file("rawdata2020.Rdata", package="SIVA"))
plot(rawdata2020$tot_vol_barrage)
# les totaliseurs sont remis à plat
cordata2020 <- traitement_siva(dat=rawdata2020)
plot(cordata2020$tot_vol_barrage)
}
