% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trevelo.R
\docType{data}
\name{trevelo}
\alias{trevelo}
\title{Variables pour le trevelo}
\format{
\subsection{\code{variable}}{

Un data frame avec 9 lignes et 9 colonnes:
\describe{
\item{code}{Code interne pour appel de la variable}
\item{tag}{Le tag}
\item{tagstation}{Le tag de la station}
\item{libelle}{Le libelle modifié depuis SIVA}
\item{unite}{L'unite}
\item{tablehisto}{Le nom de la table dans la base}
}
}
}
\usage{
trevelo
}
\description{
extrait depuis SIVA_R voir data_raw variable.R
}
\keyword{datasets}
