% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_vega.R
\name{update_vega}
\alias{update_vega}
\title{Mets à jour les dernières données depuis l'interface REST de la sonde VEGA}
\usage{
update_vega(con, sonde = "Trevelo", apiKey)
}
\arguments{
\item{con}{Une connexion pool vers SIVA}

\item{sonde}{"Trevelo" ou "Isac"}

\item{apiKey}{La clé pour se connecter à l'interface rest de Vega}
}
\description{
Récupère la date de la dernière donnée puis mets à jour la table correspondant
à Trevelo ou Isac
}
\examples{
\dontrun{
update_vega(con = pool, sonde = "Trevelo", apiKey=apikey)
update_vega(con= pool, sonde = "Isac", apiKey=apikey)
}
}
