% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\docType{data}
\name{variable}
\alias{variable}
\title{Variables de SIVA}
\format{
\subsection{\code{variable}}{

A data frame with 878 rows and 18 columns:
\describe{
\item{Tag}{le tag}
\item{TagStation}{Le tag de la station}
\item{Libelle}{le libelle}
...
}
}
}
\usage{
variable
}
\description{
extrait depuis SIVA_R voir data_raw variable.R
}
\keyword{datasets}
