% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilansiva-class.R
\docType{class}
\name{bilansiva-class}
\alias{bilansiva-class}
\title{Class BilanSiva
Cette classe utilise TableSiva pour récupérer les infos sur plusieurs tables}
\description{
Class BilanSiva
Cette classe utilise TableSiva pour récupérer les infos sur plusieurs tables
}
\section{Slots}{

\describe{
\item{\code{table}}{= noms des la tables a aller chercher}

\item{\code{nom}}{= noms des la colonnes}

\item{\code{daterondes}}{= traitement par la methode datesrondes, soit "constant" pour ouvertures en tout ou rien des vannes, soit "linear" pour des données continues, soit "none" pour pas de traitement}

\item{\code{tags}}{= si plusieurs variables sont dans la même table il faut donner leur tag pour les différentier}

\item{\code{debut}}{= date de début}

\item{\code{fin}}{= date de fin}
}}

\note{
Cette classe teste que les longueurs des champs nom, daterondes,et table sont égales
}
