% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debit_total.R
\name{debit_total}
\alias{debit_total}
\title{Débit de l'ensemble des volets et des vannes du barrage.}
\usage{
debit_total(
  param = c(Cvo = 0.392, Cvg = 1.5, Cvgs = 1.5, Cvw = 1.5),
  param0,
  dat
)
}
\arguments{
\item{param}{Les paramètres à optimiser (4 paramètres "Cvo","Cvg","Cvgs","Cvw") voir}

\item{param0}{Paramètre sans optim}

\item{dat}{Tableau de données formaté en entrée, colonnes volet1,volet2,volet3,
volet4,volet5,vanne1,vanne2,vanne3,vanne4,vanne5,debit_vilaine_estime,debit_passe,
debit_moyen_cran,tot_vol_vanne,tot_vol_passe,tot_vol_siphon,tot_vol_volet,
tot_vol_ecluse,niveauvilaine,niveaumer,niveauvilaineb,niveaumerb,debit_siphon_1,
debit_siphon_2,debit_vanne1,debit_vanne2,debit_vanne3,debit_vanne4,debit_vanne5,
debit_volet1,debit_volet2,debit_volet3,debit_volet4,debit_volet5,tot_vol}
}
\value{
Un data frame avec
\item{date}{la date (au format date),}
\item{Qva12345}{la somme des débits des vannes 1 à 5,}
\item{Qvo12345}{la somme des débits des volets 1 à 5,}
\item{tot_vol_vanne}{Le volume total évacué par les vannes,}
\item{tot_vol_ecluse}{le volume total évacué par l'écluse,}
\item{tot_vol_siphon}{le volume total évacucé par les deux siphons,}
\item{tot_vol_passe}{le volume total évacucé par la passe,}
\item{debit_moyen_cran}{le débit moyen journalier au pont de Cran}
}
\description{
Débit total (cette fonction retourne le débit total journalier et la valeur moyenne
Pondérée du débit de Cran.) en utilisant les choix IFSW pour les fonctions de débit
Briand et Woimant (2015). Le calcul se fait en utilisant canal="ifsw" et loi_orificenoye = "ifws".
}
\note{
Cette fonction a été utilisée pour optimiser les coefficients de débit,
param est consititué des coefficients optimisés
mais il faut les autres qui sont stocké dans param0 pour faire tourner débit_total
Ainsi a_optimiser tourne bien sur param mais débit total sur param + les coeff de param0.
CHANGEMENT 2016 la référence devient niveaumerb et niveauVilaineb.
}
