% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debit_vannes.R
\name{debit_vannes}
\alias{debit_vannes}
\title{débit de l'ensemble des vannes du barrage d'Arzal}
\usage{
debit_vannes(
  horodate,
  hvanne1,
  hvanne2,
  hvanne3,
  hvanne4,
  hvanne5,
  hvilaine,
  hmer,
  canal = "manning",
  Cma = 0.65,
  Cho = 0.79,
  m_oninf4 = 0.663,
  m_onsup4 = 0.995,
  Cvg = 1.5,
  Cvgs = 1.5,
  Cvw = 1.5,
  loi_orificenoye = "ifws"
)
}
\description{
Le débit de toutes les vannes est calculé à partir de la fonction de calcul du débit sur une vanne
et des coefficients de débit
}
\examples{
# chargement des paramètres
load(system.file("param2012_2014.Rdata", package = "SIVA"))
# les valeurs ajustées des paramètres sont
knitr::kable(param)
load(system.file("dat2019.Rdata", package = "SIVA"))
head(dat)

}
