% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debit_volets.R
\name{debit_volets}
\alias{debit_volets}
\title{Calcul du débit sur tous les volets}
\usage{
debit_volets(
  hvolet1,
  hvolet2,
  hvolet3,
  hvolet4,
  hvolet5,
  hvilaine,
  hvanne1,
  hvanne2,
  hvanne3,
  hvanne4,
  hvanne5,
  hmer,
  Cvo = 0.627
)
}
\arguments{
\item{hvolet1}{La hauteur du volet1.}

\item{hvolet2}{La hauteur du volet2.}

\item{hvolet3}{La hauteur du volet3.}

\item{hvolet4}{La hauteur du volet4.}

\item{hvolet5}{La hauteur du volet5.}

\item{hvilaine}{Le niveau Vilaine.}

\item{hvanne1}{La hauteur de vanne.}

\item{hvanne2}{La hauteur de vanne.}

\item{hvanne3}{La hauteur de vanne.}

\item{hvanne4}{La hauteur de vanne.}

\item{hvanne5}{La hauteur de vanne.}

\item{hmer}{Le niveau mer.}

\item{Cvo}{Le coefficient de calcul de débit sur les volets, défaut 0.627.}
}
\value{
Un data frame avec les colonnes Qvolet1 ... Qvolet5.
}
\description{
Somme du débit de tous les volets
}
