% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicateur_brochet_isac.R
\name{indicateur_brochet_isac}
\alias{indicateur_brochet_isac}
\title{Analyse de l'indicateur brochet sur l'Isac}
\usage{
indicateur_brochet_isac(
  dat,
  h_favorable = 2.3,
  h_moyen = 2.2,
  debut_ponte = "25/02",
  fin_ponte = "15/03",
  fin_emergence = "01/05",
  niveau_marais = "isac_niveau_marais"
)
}
\arguments{
\item{dat}{Un tableau de données, chargé par load_isac et traitement_debit_isac.}

\item{h_favorable}{Le niveau favorable pour la reproduction.}

\item{h_moyen}{Hauteur en dessus de laquelle la reproduction est considérée comme
moyenne et et en dessous duquel elle est considérée comme mauvaise.}

\item{debut_ponte}{Date de début de ponte defaut "25/02"}

\item{fin_ponte}{Date de fin de ponte et début de l'émergence defaut "15/03"}

\item{fin_emergence}{La date de fin de l'emergence (exclue) defaut "01/05".}

\item{niveau_marais}{Le nom de la variable de niveau dans les marais, defaut "isac_niveau_marais"
voir le jeu de données isac pour les codes possibles.}
}
\value{
Ub vecteur (factor) des types pour le brochet
}
\description{
note :si on est pas dans la bonne période, pas de calcul du niveau ponte, on
prend 2.30 comme référence....)
}
