% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_debit_barrage.R
\name{load_debit_barrage}
\alias{load_debit_barrage}
\title{Chargement des donnees de debit du barrage}
\usage{
load_debit_barrage(debut, fin, con)
}
\arguments{
\item{debut}{La date de début format POSIXct ou character "\%Y-\%m-\%d \%H:\%M:\%S}

\item{fin}{La fin format POSIXct ou character "\%Y-\%m-\%d \%H:\%M:\%S}

\item{con}{une connexion pool}
}
\value{
Le tableau de données pour le calcul des debits du barrage
hauteur :
"volet1","volet2","volet3","volet4","volet5",
"vanne1","vanne2","vanne3","vanne4","vanne5",
débits :
"debit_vilaine_estime (b_barrage_debit(2515))",
"debit_passe",
"debit_moyen_cran",
totaliseurs volumes :
"tot_vol_vanne","tot_vol_volet","tot_vol_ecluse",
les variables tot_vol_passe et tot_vol_siphon ne sont plus chargées
car fausses. Elles seront recalculées par traitement_siva
niveaux :
"niveauvilaine"(passe),"niveaumer"(passe),"niveauvilaineb","niveaumerb",
débits barrage :
"debit_siphon_1","debit_siphon_2",
"debit_vanne1","debit_vanne2","debit_vanne3","debit_vanne4","debit_vanne5",
"debit_volet1","debit_volet2","debit_volet3","debit_volet4","debit_volet5"
}
\description{
Il faut avoir configuré la connexion à la base
}
\examples{
\dontrun{
load_debit_barrage(debut = 
as.POSIXct(strptime("2010-01-01 00:00:00", format = "\%Y-\%m-\%d \%H:\%M:\%S")),
   fin = 
   as.POSIXct(strptime("2010-01-10 00:00:00", format = "\%Y-\%m-\%d \%H:\%M:\%S")),
   con= con)
   }
}
