% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_niveaux.R
\name{load_niveaux}
\alias{load_niveaux}
\title{Chargement des donnees de niveaux}
\usage{
load_niveaux(
  debut,
  fin,
  redon = TRUE,
  painfault = TRUE,
  langon = TRUE,
  malon = TRUE,
  aucfer = TRUE,
  molac = TRUE,
  potinais = TRUE,
  leguelin = TRUE,
  Sixt = TRUE
)
}
\arguments{
\item{debut}{La date de début format POSIXct}

\item{fin}{La fin}
}
\value{
Le tableau de données de niveaux
}
\description{
Il faut avoir configuré la connexion à la base
}
\examples{
if (interactive){
load_debit_barrage(debut = as.POSIXct(strptime("2010-01-01 00:00:00", 
format = "\%Y-\%m-\%d \%H:\%M:\%S")),
   fin = as.POSIXct(strptime("2010-01-10 00:00:00", 
   format = "\%Y-\%m-\%d \%H:\%M:\%S")))
   }
}
