% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddb-bilansiva.R
\name{loaddb,bilansiva-method}
\alias{loaddb,bilansiva-method}
\title{Methode de chargement pour loadb}
\usage{
\S4method{loaddb}{bilansiva}(objet, datawd, graphe = TRUE, restart = NULL)
}
\arguments{
\item{objet}{Un objet de class bilansiva.}

\item{datawd}{Le chemin vers le repertoire ou sauver les données.}

\item{graphe}{Boolean, faut il un graphe ?}

\item{restart}{Default NULL, si non null, l'objet est sauvé sous "tempsauvobjet" dans datawd.}
}
\value{
Un objet de classe bilansiva
}
\description{
Methode de chargement pour loadb
}
