% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddb-bilansiva.R
\name{loaddb,bilansiva-method}
\alias{loaddb,bilansiva-method}
\title{Methode de chargement pour loadb}
\usage{
\S4method{loaddb}{bilansiva}(
  objet,
  datawd,
  con,
  checktag = FALSE,
  plot = TRUE,
  loadrdata = FALSE,
  saverdata = FALSE
)
}
\arguments{
\item{objet}{Un objet de class bilansiva.}

\item{datawd}{Le chemin vers le repertoire où sauver les données.}

\item{con}{Une connexion pool}

\item{checktag}{Defaut FALSE, faut il vérifier que le tag existe dans la table,
à utiliser lors des tests.}

\item{plot}{Boolean, faut il un graphe ?}

\item{loadrdata}{Faut il charger un jeu de donnees sauvé dans datawd à la place de la base}

\item{saverdata}{Faut il sauvegarder un jeu de donnees dans datawd}
}
\value{
Un objet de classe bilansiva
}
\description{
Methode de chargement pour loadb
}
\examples{
\dontrun{
pool <- pool::dbPool(
   drv = RMariaDB::MariaDB(),
   dbname = "archive_IAV",
   host = hostmysql.,
   username = umysql.,
   password = pwdmysql.
)
bil<-new("bilansiva",
        tables=c("b_langon_debit", 
                 "b_pont_de_cran_debit"
        ),
        noms=c(
          "debit_langon",
          "debit_moyen_cran"          
        ),
        tags=as.integer(c(3000,1900)),
        daterondes=rep("constant",2), 
        debut=as.POSIXct(strptime(paste0(2019,"-05-01 00:00:00"),
                                  format="\%Y-\%m-\%d \%H:\%M:\%S")),
        fin=as.POSIXct(strptime(paste0(2019,"-09-01 00:00:00"),
                                format="\%Y-\%m-\%d \%H:\%M:\%S"))
)
res <- loaddb(bil, con = pool)
poolClose(pool)
}
}
