% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddb-tablesiva.R
\name{loaddb,tablesiva-method}
\alias{loaddb,tablesiva-method}
\title{Methode loaddb}
\usage{
\S4method{loaddb}{tablesiva}(objet)
}
\arguments{
\item{objet, }{un objet de classe tablesiva}
}
\value{
Un object de classe tablesiva
}
\description{
La méthode loaddb permet d'aller chercher les données après avoir installé le driver odbc et avoir appellé la bonne chaine de connection (il doit y avoir une variable nommée sivacon dans le workspace)
}
\examples{
# sur mon ordi j'ai des mots de passe chargés au démarrage à partir de Rprofile.site
# Ici test si existent et si oui il faut le main password pour les decrypter, sinon il
# faut les entrer après un prompt du programme. 
if (interactive()){
if (!exists("mainpass")) mainpass <- getPass::getPass(msg = "main password")
if (!exists("hostmysql")) {
  hostmysql. <- getPass::getPass(msg = "Saisir host")
  # ci dessous pour ne pas redemander au prochain tour
  hostmysql <- encrypt_string(string = hostmysql., key = mainpass)
} else {
  hostmysql. <- decrypt_string(string = hostmysql, key = mainpass)
}
if (!exists("pwdmysql")) {
  pwdmysql. <- getPass::getPass(msg = "Saisir password")
  pwdmysql <- encrypt_string(string = pwdmysql., key = mainpass)
}  else {
  # pass should be loaded
  pwdmysql. <- decrypt_string(string = pwdmysql, key = mainpass)
}
if (!exists("umysql")) {
  umysql. <- getPass::getPass(msg = "Saisir user")
  umysql <- encrypt_string(string = umysql., key = mainpass)
} else {
  umysql. <- decrypt_string(string = umysql, key = mainpass)
}

options(stacomiR.dbname="archive_IAV", # TODO not used....
        stacomiR.host = hostmysql.,
        stacomiR.password = pwdmysql.,
        stacomiR.user = umysql.,
        stacomiR.ODBClink = "archive_IAV")

tablesiva <-
  new(
    "tablesiva",
    debut = as.POSIXct(as.Date("2021-10-25")),
    fin =  as.POSIXct(as.Date("2021-12-01")),
    table = "b_barrage_volet4_hauteur",
    nom = "volet4"
  )
loaddb(tablesiva)
}
}
