% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddb-tablesiva.R
\name{loaddb,tablesiva-method}
\alias{loaddb,tablesiva-method}
\title{Methode loaddb}
\usage{
\S4method{loaddb}{tablesiva}(objet, con)
}
\arguments{
\item{objet, }{un objet de classe tablesiva}

\item{con, }{une connexion pool}
}
\value{
Un object de classe tablesiva
}
\description{
La méthode loaddb permet d'aller chercher les données après avoir installé le driver odbc et avoir appellé la bonne chaine de connection (il doit y avoir une variable nommée sivacon dans le workspace)
}
