% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_journalier_vanne_volet.R
\name{plotly_journalier_vanne_volet}
\alias{plotly_journalier_vanne_volet}
\title{Graphique du débit journalier vannes volets}
\usage{
plotly_journalier_vanne_volet(date, debit_traite, is_plotly = TRUE)
}
\arguments{
\item{date}{La date selectionnée pour le graphique de la journée.}

\item{debit_traite}{Un data frame avec les débits traités par\verb{debit total}.}

\item{is_plotly}{: défaut TRUE l'objet est transformé en graphique plotly.}
}
\value{
Un graphique ggplotly du débit journalier au barrage d'Arzal.
}
\description{
Graphique du débit journalier vannes volets
}
\examples{
\dontrun{
plotly_journalier_vanne_volet(date = "2018-01-01", debit_traite=Q12345)
}
}
