% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_volume_jour.R
\name{plotly_volume_jour}
\alias{plotly_volume_jour}
\title{graphique des volumes journaliers}
\usage{
plotly_volume_jour(QV)
}
\arguments{
\item{QV}{Un dataset des volumes journaliers (voir vignette plot)}
}
\value{
Un graphique plotly
}
\description{
graphique des volumes journaliers
}
\examples{
\dontrun{
plotly_volume_jour(QV=QV)
}
}
