% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivaplot-tablesiva.R
\name{sivaplot,tablesiva-method}
\alias{sivaplot,tablesiva-method}
\title{Methode sivaplot}
\usage{
\S4method{sivaplot}{tablesiva}(objet)
}
\arguments{
\item{objet, }{un objet de classe tablesiva}
}
\value{
Un object de classe tablesiva
}
\description{
La méthode sivaplot permet d'aller chercher les données après avoir installé le driver odbc et avoir appellé la bonne chaine de connection (il doit y avoir une variable nommée sivacon dans le workspace)
}
