% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablesiva.R
\docType{class}
\name{Tablesiva-class}
\alias{Tablesiva-class}
\title{Classe de connexion SIVA}
\value{
Un objet de classe Tablesiva.
}
\description{
Classe de connexion SIVA
}
\section{Slots}{

\describe{
\item{\code{table,}}{Nom de la table (serie de données) dans SIVA.}

\item{\code{nom,}}{Nom de la série dans le tableau de données.}

\item{\code{tag,}}{Tag dans siva}

\item{\code{debut,}}{horodate de début.}

\item{\code{fin,}}{horodate de fin.}

\item{\code{rawdata,}}{données brutes.}

\item{\code{corrdata,}}{données corrigées.}
}}

