% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement_siva.R
\name{traitement_siva}
\alias{traitement_siva}
\title{Traitement des données en sortie de SIVA}
\usage{
traitement_siva(dat)
}
\arguments{
\item{dat}{Un tableau de données contenant des données totalisées
par exemple chargées par la fonction \code{load_debit_barrage}}
}
\value{
Un tableau de données avec diff pour les totaux,
tot_vol_vanne, tot_vol_passe, tot_vol_siphon, tot_vol_volet, les débits
de siphon corrigés, et une colonne tot_vol= volume total journalier
au barrage.
}
\description{
recalcule des données des compteurs, enlève les valeurs abérrantes des débits
des siphons. Les données doivent avoir 'tot_....' en entête, toutes, les colonnes
contenant 'tot' font l'objet d'une différence entre les valeurs des lignes 2:n et les
lignes 1:n-1, les données de la première ligne sont NA
Pour les données des Siphons debit_siphon_1, et debit_siphon_2 les débits > 3.8
sont ramenés à 3.8
}
\examples{
# voir example-bilansiva-debit pour le chargement des données de 2020 dans SIVA
load(system.file("rawdata2020.Rdata", package="SIVA"))
plot(rawdata2020$tot_vol_vanne)
# les totaliseurs sont remis à plat
cordata2020 <- traitement_siva(dat=rawdata2020)
plot(cordata2020$tot_vol_vanne)
}
